Shark Game Init Plus Gameplay
Designed and implemented a brief shark game, where the shark is controlled by rotating it up and down to find the glowing fish to eat.

Glowing fish are chosen from a list of fish in the scene which are randomly placed in the scene. The game can be reset after 3 "bad" fish are eaten or 5 minutes have passed.